fix(v3.12.0 baseline): clippy clean + benches auto_increment + fmt
- benches/tpch_comprehensive.rs (14) and benches/bench_insert.rs (4)
  and benches/dataset_generator.rs (6) and benches/qps_bench.rs (3)
  and benches/bench_insert.rs (4) — add `auto_increment: false` to
  ColumnDefinition initializers (PR #4365 added the field but didn't
  sync bench callers; cargo bench --no-run failed).
- cargo fmt --all (166 files, mostly whitespace drift on rustc 1.97).
- Add `#![allow(dead_code)]` to 16 shared test-infra files (multi-TU
  harnesses whose helpers are exercised per-test by some TUs only).
- Manual clippy cleanups:
  - empty-line-after-doc-comment: tests/common/mod.rs:487 (referenced
    from 3 paths), tests/integration/tpch/tpch_sf1_22_vs_3engines_test.rs:68,
    tests/operators/date_string_compare.rs:8
  - unused_must_use: tests/integration/sql/q21_cell_regression_test.rs:76,82
  - manual checked division → checked_div: parallel_executor_test.rs,
    parallel_executor_harness.rs, performance_schema_test.rs
  - ptr_arg (&PathBuf → &Path): four_way_compare_test.rs:130,191 and
    tpch_22_mysql_cli_wire_test.rs:65,191
  - redundant pattern: mysql_wire_protocol_test.rs:52,
    mysqladmin_e2e_test.rs:85
  - unwrap-after-is_some: tpch_q9_audit.rs:116
  - unwrap_err-after-is_ok (3): tx_wal_contract_tests.rs WAL-002/004/005
  - assert!(true) no-op stubs (5): tx_wal_contract_tests.rs WAL-001/006/007/008/009
  - assert!(false) → panic!: upsert_test.rs:92 (unimplemented feature)
  - useless compare on u64: server_health_test.rs:42,69
  - inconsistent digit grouping: q4_hash_semi_join_test.rs:161
  - duplicated #[test]: pr_template_and_full_gate_test.rs:104/105
  - manual `0..n` counter → `(1_u64..).take(n)`: recovery_fuzzer_test.rs:646
  - unused vars `attempted_txns`, `completed_txns` → underscore-prefix:
    crash_monkey_test.rs (test counter, kept for telemetry)
  - useless use of format!: arch_sem_debt_gate_test.rs
  - very complex type → type alias: tpch_bench.rs:128
  - unused doc comment over macro invocation: qps_bench.rs:92
  - modulo_one (`round % 1` always 0): src/execution_engine_tests.rs:983

Verified:
- cargo check --all-features: PASS
- cargo fmt --check --all: PASS
- cargo clippy --all-features --all-targets -- -D warnings: EXIT 0
- cargo bench --no-run --all-features: 0 errors (was: 18 sites E0063)
- cargo test --all-features --no-run: all binaries compile

Generated artifacts (coverage-report/, docs/releases/v3.7.0/html/,
docs/audit/status/2026-06-07-tpch-cell-diff-v390.json, data/soak-*/*.wal,
docs/formal/PROOF_023_*, smoke-report.md) intentionally left unstaged —
these are test/CI output, not source changes.

+++ benches/common.rs
@@ -180,7 +180,7 @@ fn native_password_auth(password: &[u8], scramble: &[u8; SCRAMBLE_LEN]) -> [u8; 
 fn build_handshake_response41(user: &str, auth_response: &[u8]) -> wire_err::Result<Vec<u8>> {
     let mut p = Vec::with_capacity(64 + user.len() + auth_response.len());
 
-    p.extend_from_slice(&(CLIENT_CAPABILITIES as u32).to_le_bytes());
+    p.extend_from_slice(&CLIENT_CAPABILITIES.to_le_bytes());
     p.extend_from_slice(&MAX_PACKET_SIZE.to_le_bytes());
     p.push(CHARSET_UTF8);
     p.extend_from_slice(&[0u8; 23]); // 23 reserved bytes
